Composer downloads very old 1.7.x-dev 8f77f13
Looks like "1.7.*" (as well as "~1.7", which I also tried on another fork) erroneously downloads very old 1.7.x-dev 8f77f13. It leads to lots of errors as old version missed lots of fixes. That's not intended I guess.
This fix make it to download current dev version. That's not very good too (stable versions should be preferred), but it is affordable.
If someone could imagine a way to make it to work as desired - you're welcome.